decorate column highlight function to run in bg
Blind debugging the app which was crashing Pythonista, I found that starting the loop would sometimes cause a crash.  I notice that this function wasn't decorated to run in the background, as the function for the button action is.  I added it and it seems to help when tested under macOS.